fix tls for ntlm
