--- doc.go
@@ -1,8 +1,0 @@
-/*
-
-Package wrapping Exchange Web Service (EWS)
-
-https://docs.microsoft.com/en-us/exchange/client-developer/web-service-reference/ews-operations-in-exchange
-
-*/
-package ews
