--- ewsutil/doc.go
@@ -1,2 +1,0 @@
-// Utility package to contain functions like sendMail, create Calendar appointment and get user availability
-package ewsutil
