+++ fault_test.go
@@ -2,12 +2,13 @@ package ews
 
 import (
 	"errors"
-	"github.com/stretchr/testify/assert"
 	"io"
 	"net/http"
 	"reflect"
 	"strings"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 const (
