+++ get_room_lists.go
@@ -1,7 +1,5 @@
 package ews
 
-import "encoding/xml"
-
 type GetRoomListsRequest struct {
 	XMLName struct{} `xml:"m:GetRoomLists"`
 }
@@ -26,25 +24,4 @@ type getRoomListsResponseEnvelop struct {
 }
 type getRoomListsResponseBody struct {
 	GetRoomListsResponse GetRoomListsResponse `xml:"GetRoomListsResponse"`
-}
-
-func GetRoomLists(c Client) (*GetRoomListsResponse, error) {
-
-	xmlBytes, err := xml.MarshalIndent(&GetRoomListsRequest{}, "", "  ")
-	if err != nil {
-		return nil, err
-	}
-
-	bb, err := c.SendAndReceive(xmlBytes)
-	if err != nil {
-		return nil, err
-	}
-
-	var soapResp getRoomListsResponseEnvelop
-	err = xml.Unmarshal(bb, &soapResp)
-	if err != nil {
-		return nil, err
-	}
-
-	return &soapResp.Body.GetRoomListsResponse, nil
 }
