+++ go.mod
@@ -1,10 +1,15 @@
-module github.com/mhewedy/ews
+module github.com/gunsluo/goews
 
-go 1.12
+go 1.18
 
 require (
-	github.com/Azure/go-ntlmssp v0.0.0-20191115210519-2b2be6cc8ed4
-	github.com/stretchr/testify v1.4.0
-	golang.org/x/crypto v0.0.0-20191206172530-e9b2fee46413 // indirect
-	golang.org/x/lint v0.0.0-20191125180803-fdd1cda4f05f // indirect
+	github.com/Azure/go-ntlmssp v0.0.0-20211209120228-48547f28849e
+	github.com/stretchr/testify v1.7.2
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/crypto v0.0.0-20220525230936-793ad666bf5e // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
